revert freetype-2.8-2.4.11-API
This reverts commit 8eb61d16dd5c24b1025e6f978734e2a5e4a71274.

+++ src/pcf/pcfdrivr.c
@@ -523,10 +523,10 @@ THE SOFTWARE.
 
     metric = face->metrics + glyph_index;
 
-    bitmap->rows       = ( metric->ascent +
-                           metric->descent );
-    bitmap->width      = ( metric->rightSideBearing -
-                           metric->leftSideBearing );
+    bitmap->rows       = (unsigned int)( metric->ascent +
+                                         metric->descent );
+    bitmap->width      = (unsigned int)( metric->rightSideBearing -
+                                         metric->leftSideBearing );
     bitmap->num_grays  = 1;
     bitmap->pixel_mode = FT_PIXEL_MODE_MONO;
